do not move image to res folder if already exists

+++ sort_images.py
@@ -8,7 +8,6 @@ from settings import Settings
 
 
 def run_image_sorter(args: Settings):
-
     pathes = glob.glob(args.samples_dir + "*")
     for path in pathes:
         if os.path.isdir(path):
@@ -19,19 +18,22 @@ def run_image_sorter(args: Settings):
             x, y = image.size
             dir_name = f"{x}x{y}"
             image.close()
-            move_file(args.samples_dir, dir_name, path)
+            maybe_move_file(args.samples_dir, dir_name, path)
             print(f"Moved: {dir_name}: {path}")
 
         except UnidentifiedImageError:
             dir_name = "bad"
-            move_file(args.samples_dir, dir_name, path)
+            maybe_move_file(args.samples_dir, dir_name, path)
             print(f"Moved {dir_name}: {path}")
             pass
 
 
-def move_file(samples_dir: str, dir_name: str, path: str):
-
-    dir_path = samples_dir + dir_name
+def maybe_move_file(samples_dir: str, dir_name: str, path: str):
+    dir_path = os.path.join(samples_dir, dir_name)
+    file_name = os.path.basename(path)
+    file_path = os.path.join(dir_path, file_name)
+    if os.path.exists(file_path):
+        return
     if not os.path.exists(dir_path):
         os.makedirs(dir_path)
     shutil.move(path, dir_path)
